Move unique_ptr to library
