+++ src/tbb/scheduler_common.h
@@ -41,6 +41,7 @@
 
 #include <cstdint>
 #include <exception>
+#include <memory> // unique_ptr
 
 //! Mutex type for global locks in the scheduler
 using scheduler_mutex_type = __TBB_SCHEDULER_MUTEX_TYPE;
@@ -67,6 +68,22 @@ template<task_stream_accessor_type> class task_stream;
 
 using isolation_type = std::intptr_t;
 constexpr isolation_type no_isolation = 0;
+
+struct cache_aligned_deleter {
+    template <typename T>
+    void operator() (T* ptr) const {
+        ptr->~T();
+        cache_aligned_deallocate(ptr);
+    }
+};
+
+template <typename T>
+using cache_aligned_unique_ptr = std::unique_ptr<T, cache_aligned_deleter>;
+
+template <typename T, typename ...Args>
+cache_aligned_unique_ptr<T> make_cache_aligned_unique(Args&& ...args) {
+    return cache_aligned_unique_ptr<T>(new (r1::cache_aligned_allocate(sizeof(T))) T(std::forward<Args>(args)...));
+}
 
 //------------------------------------------------------------------------
 // Extended execute data
